cleanup the code a little to reduce compiler warnings
git-svn-id: svn://svn.code.sf.net/p/cxxtools/code/trunk@1161 15ae5fad-cc11-0410-8fac-bce609e504b0

+++ cxxtools/src/query_params.cpp
@@ -197,8 +197,8 @@ static void appendUrl(std::string& url, char ch)
     url += '%';
     char hi = (ch >> 4) & 0x0f;
     char lo = ch & 0x0f;
-    url += hex[hi];
-    url += hex[lo];
+    url += hex[static_cast<int>(hi)];
+    url += hex[static_cast<int>(lo)];
   }
 }
